+++ packages/vite-app-ts/.eslintrc
@@ -29,7 +29,6 @@
     "plugin:import/warnings",
     "plugin:import/typescript",
     "plugin:react-hooks/recommended",
-    "plugin:jest/recommended",
     "plugin:react/recommended",
     "plugin:tailwind/recommended",
     "plugin:prettier/recommended",
@@ -37,7 +36,7 @@
     "plugin:@typescript-eslint/recommended",
     "plugin:@typescript-eslint/recommended-requiring-type-checking"
   ],
-  "plugins": ["prettier", "react", "react-hooks", "unused-imports", "testing-library"],
+  "plugins": ["prettier", "react", "react-hooks", "unused-imports"],
   "rules": {
     "eqeqeq": ["error", "smart"],
     "react/jsx-wrap-multilines": "warn",
