+++ packages/vite-app-ts/package.json
@@ -38,8 +38,8 @@
     "authereum": "^0.1.14",
     "axios": "^0.23.0",
     "buffer": "^6.0.3",
-    "eth-components": "2.2.1",
-    "eth-hooks": "3.4.0",
+    "eth-components": "^3.3.1",
+    "eth-hooks": "^4.3.0",
     "ethereum-cryptography": "^0.1.3",
     "ethers": "^5.4.1",
     "fortmatic": "^2.2.1",
@@ -47,6 +47,7 @@
     "graphql": "^15.6.1",
     "history": "^5.0.0",
     "http-browserify": "^1.7.0",
+    "invariant": "^2.2.4",
     "pretty-time": "^1.1.0",
     "process": "^0.11.10",
     "qrcode.react": "^1.0.1",
@@ -65,6 +66,7 @@
     "timers-browserify": "^2.0.12",
     "tiny-invariant": "^1.1.0",
     "use-debounce": "^7.0.0",
+    "usehooks-ts": "^2.5.4",
     "walletlink": "^2.1.11",
     "web3modal": "^1.9.1"
   },
