+++ packages/vite-app-ts/src/app/routes/main/hooks/useScaffoldAppProviders.ts
@@ -1,9 +1,10 @@
 import { StaticJsonRpcProvider } from '@ethersproject/providers';
+import { EthersModalConnector, useEthersAppContext } from 'eth-hooks/context';
 import { TEthersProvider, TNetworkInfo } from 'eth-hooks/models';
-import { useCallback, useEffect, useMemo, useState } from 'react';
-import { ICoreOptions } from 'web3modal';
-import { EthersModalConnector, useEthersContext, CreateEthersModalConnector } from 'eth-hooks/context';
+import { useCallback, useEffect, useState } from 'react';
 import { useThemeSwitcher } from 'react-css-theme-switcher';
+import { ICoreOptions } from 'web3modal';
+
 import { mainnetProvider, localProvider, targetNetworkInfo } from '~~/config/providersConfig';
 
 export interface IScaffoldAppProviders {
@@ -11,12 +12,12 @@ export interface IScaffoldAppProviders {
   targetNetwork: TNetworkInfo;
   mainnetProvider: StaticJsonRpcProvider;
   localProvider: StaticJsonRpcProvider;
-  createLoginConnector: CreateEthersModalConnector;
+  createLoginConnector: any;
 }
 
 export const useScaffoldProviders = (): IScaffoldAppProviders => {
   const [web3Config, setWeb3Config] = useState<Partial<ICoreOptions>>();
-  const ethersContext = useEthersContext();
+  const ethersContext = useEthersAppContext();
 
   useEffect(() => {
     // import async to split bundles
@@ -30,7 +31,7 @@ export const useScaffoldProviders = (): IScaffoldAppProviders => {
 
   const { currentTheme } = useThemeSwitcher();
 
-  const createLoginConnector: CreateEthersModalConnector = useCallback(
+  const createLoginConnector: any = useCallback(
     (id?: string) => {
       if (web3Config) {
         const connector = new EthersModalConnector(
@@ -52,7 +53,7 @@ export const useScaffoldProviders = (): IScaffoldAppProviders => {
   }, [web3Config]);
 
   return {
-    currentProvider: ethersContext.ethersProvider ?? localProvider,
+    currentProvider: ethersContext.provider ?? localProvider,
     mainnetProvider: mainnetProvider,
     localProvider: localProvider,
     targetNetwork: targetNetworkInfo,
